♻️ refactor(monitoring): extract data model selectors
Move monitoring types, range helpers, auth metadata, API key display logic, and row builders into model modules.

Keep the useMonitoringData public exports stable while reducing the hook to data orchestration and remaining analytics adapters.
