+++ src/features/monitoring/model/base.ts
@@ -1,0 +1,75 @@
+export type RecordLike = Record<string, unknown>;
+
+export const isRecord = (value: unknown): value is RecordLike =>
+  value !== null && typeof value === 'object' && !Array.isArray(value);
+
+export const readString = (value: unknown) => {
+  if (value === null || value === undefined) return '';
+  const text = String(value).trim();
+  return text;
+};
+
+export const parseBoolean = (value: unknown) => {
+  if (typeof value === 'boolean') return value;
+  if (typeof value === 'number') return value !== 0;
+  if (typeof value === 'string') {
+    const normalized = value.trim().toLowerCase();
+    return ['1', 'true', 'yes', 'on'].includes(normalized);
+  }
+  return false;
+};
+
+export const extractArrayPayload = (payload: unknown, key: string): unknown[] => {
+  if (Array.isArray(payload)) return payload;
+  if (!isRecord(payload)) return [];
+  const candidate = payload[key] ?? payload.items ?? payload.data ?? payload;
+  return Array.isArray(candidate) ? candidate : [];
+};
+
+export const extractHost = (baseUrl: string) => {
+  const trimmed = readString(baseUrl);
+  if (!trimmed) return '-';
+
+  try {
+    return new URL(trimmed).host || trimmed;
+  } catch {
+    return trimmed.replace(/^https?:\/\//i, '').split('/')[0] || trimmed;
+  }
+};
+
+export const joinUnique = (values: Iterable<string>, limit = 3) => {
+  const unique = Array.from(
+    new Set(
+      Array.from(values)
+        .map((value) => value.trim())
+        .filter(Boolean)
+    )
+  );
+  if (unique.length <= limit) {
+    return unique.join(', ');
+  }
+  return `${unique.slice(0, limit).join(', ')} +${unique.length - limit}`;
+};
+
+export const maskEmailLike = (value: string) => {
+  const trimmed = value.trim();
+  const match = trimmed.match(/^([^@\s]{1,3})[^@\s]*@(.+)$/);
+  if (!match) return trimmed;
+  return `${match[1]}***@${match[2]}`;
+};
+
+export const maskAuthIndex = (value: string) => {
+  const trimmed = value.trim();
+  if (!trimmed || trimmed === '-') return '-';
+  if (trimmed.length <= 10) return trimmed;
+  return `${trimmed.slice(0, 4)}...${trimmed.slice(-4)}`;
+};
+
+export const buildSearchText = (...parts: Array<string | number | boolean | null | undefined>) =>
+  parts
+    .map((part) => (part === null || part === undefined ? '' : String(part).trim().toLowerCase()))
+    .filter(Boolean)
+    .join(' ');
+
+export const formatApiKeyHashLabel = (apiKeyHash: string) =>
+  apiKeyHash ? `sha256:${apiKeyHash.slice(0, 12)}` : '-';
